requirements.txt
pytest-playwright == 0.5.1
pytest == 8.3.2
playwright == 1.45.1
allure-pytest == 2.13.5
urllib3==2.2.3
ruff==0.7.2